requirements.txt
pydantic
pyyamlannotated-types==0.7.0
pathspec==0.12.1
pydantic==2.11.2
pydantic_core==2.33.1
PyYAML==6.0.2
typing-inspection==0.4.0
typing_extensions==4.13.1
yamllint==1.37.0
